Extract capture actions
