+++ extension/platforms/chrome/hooks/useCurrentDomain.ts
