+++ extension/shared/services/platform.ts
@@ -1,7 +1,12 @@
 import type { UploadedContentFragmentTypeWithKind } from "@app/shared/lib/types";
 import type { AuthService, StoredUser } from "@app/shared/services/auth";
 import type { StorageService } from "@app/shared/services/storage";
-import type { ContentFragmentType } from "@dust-tt/client";
+import type { FileUploaderService } from "@app/ui/hooks/useFileUploaderService";
+import type {
+  ContentFragmentType,
+  ExtensionWorkspaceType,
+} from "@dust-tt/client";
+import type { ComponentType } from "react";
 
 // TODO(2025-03-19 flav): Add front platform.
 const PLATFORM_TYPES = ["chrome"] as const;
@@ -14,7 +19,14 @@ interface ConversationContext {
 export type Theme = "light" | "dark" | "system";
 export const DEFAULT_THEME: Theme = "system";
 
-export abstract class PlatformService {
+export interface CaptureActionsProps {
+  owner: ExtensionWorkspaceType;
+  isBlinking: boolean;
+  isLoading: boolean;
+  fileUploaderService: FileUploaderService;
+}
+
+export abstract class CorePlatformService {
   auth: AuthService;
   platform: PlatformType;
   storage: StorageService;
@@ -97,34 +109,9 @@ export abstract class PlatformService {
   }): Promise<void>;
 }
 
-// Shared logic that works across all platforms
-export class BasePlatformService extends PlatformService {
-  constructor(
-    platform: PlatformType,
-    authCls: new (storage: StorageService) => AuthService,
-    storage: StorageService
-  ) {
-    super(platform, authCls, storage);
-  }
-
-  // Content fragments.
-  async getFileContentFragmentId(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    conversationId: string,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    file: UploadedContentFragmentTypeWithKind
-  ): Promise<string | null> {
-    throw new Error("Platform specific implementation required.");
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async saveFilesContentFragmentIds(args: {
-    conversationId: string;
-    uploadedFiles: UploadedContentFragmentTypeWithKind[];
-    createdContentFragments: ContentFragmentType[];
-  }): Promise<void> {
-    throw new Error("Platform specific implementation required.");
-  }
+export abstract class PlatformService extends CorePlatformService {
+  // Content capture.
+  abstract getCaptureActionsComponent(): ComponentType<CaptureActionsProps>;
 }
 
 export function isValidPlatform(platform: unknown): platform is PlatformType {
