+++ extension/ui/components/conversation/AttachFragment.tsx
@@ -1,21 +1,22 @@
+import { usePlatform } from "@app/shared/context/PlatformContext";
 import { InputBarContext } from "@app/ui/components/input_bar/InputBarContext";
-import { useCurrentUrlAndDomain } from "@app/ui/hooks/useCurrentDomain";
 import type { FileUploaderService } from "@app/ui/hooks/useFileUploaderService";
 import type { ExtensionWorkspaceType } from "@dust-tt/client";
-import { Button, CameraIcon, DocumentPlusIcon } from "@dust-tt/sparkle";
 import { useContext, useEffect } from "react";
 
-type AttachFragmentProps = {
-  owner: ExtensionWorkspaceType;
+interface AttachFragmentProps {
   fileUploaderService: FileUploaderService;
   isLoading: boolean;
-};
+  owner: ExtensionWorkspaceType;
+}
 
 export const AttachFragment = ({
   owner,
   fileUploaderService,
   isLoading,
 }: AttachFragmentProps) => {
+  const platform = usePlatform();
+
   // Blinking animation.
   const { attachPageBlinking, setAttachPageBlinking } =
     useContext(InputBarContext);
@@ -29,100 +30,16 @@ export const AttachFragment = ({
     return () => clearTimeout(timer);
   }, [attachPageBlinking]);
 
-  // Blacklisting logic to disable share buttons.
-  const { currentDomain, currentUrl } = useCurrentUrlAndDomain();
-  const blacklistedConfig: string[] = owner.blacklistedDomains ?? [];
-
-  const isBlacklisted =
-    currentDomain === "chrome" ||
-    blacklistedConfig.some((d) =>
-      d.startsWith("http://") || d.startsWith("https://")
-        ? currentUrl.startsWith(d)
-        : currentDomain.endsWith(d)
-    );
+  const CaptureActionsComponent = platform.getCaptureActionsComponent();
 
   return (
-    <>
-      <div className="block sm:hidden">
-        <Button
-          icon={DocumentPlusIcon}
-          tooltip={
-            !isBlacklisted
-              ? "Attach text from page"
-              : "Attachment disabled on this website"
-          }
-          variant="outline"
-          size="sm"
-          className={attachPageBlinking ? "animate-[bgblink_200ms_3]" : ""}
-          onClick={() =>
-            fileUploaderService.uploadContentTab({
-              includeContent: true,
-              includeCapture: false,
-            })
-          }
-          disabled={isLoading || isBlacklisted}
-        />
-      </div>
-      <div className="block sm:hidden">
-        <Button
-          icon={CameraIcon}
-          tooltip={
-            !isBlacklisted
-              ? "Attach page screenshot"
-              : "Attachment disabled on this website"
-          }
-          variant="outline"
-          size="sm"
-          onClick={() =>
-            fileUploaderService.uploadContentTab({
-              includeContent: false,
-              includeCapture: true,
-            })
-          }
-          disabled={isLoading || isBlacklisted}
-        />
-      </div>
-      <div className="hidden sm:block">
-        <Button
-          icon={DocumentPlusIcon}
-          label="Add page text"
-          tooltip={
-            !isBlacklisted
-              ? "Attach text from page"
-              : "Attachment disabled on this website"
-          }
-          variant="outline"
-          size="sm"
-          className={attachPageBlinking ? "animate-[bgblink_200ms_3]" : ""}
-          onClick={() =>
-            fileUploaderService.uploadContentTab({
-              includeContent: true,
-              includeCapture: false,
-            })
-          }
-          disabled={isLoading || isBlacklisted}
-        />
-      </div>
-      <div className="hidden sm:block">
-        <Button
-          icon={CameraIcon}
-          label="Add page screenshot"
-          tooltip={
-            !isBlacklisted
-              ? "Attach page screenshot"
-              : "Attachment disabled on this website"
-          }
-          variant="outline"
-          size="sm"
-          onClick={() =>
-            fileUploaderService.uploadContentTab({
-              includeContent: false,
-              includeCapture: true,
-            })
-          }
-          disabled={isLoading || isBlacklisted}
-        />
-      </div>
-    </>
+    <div className="flex flex-row gap-2">
+      <CaptureActionsComponent
+        fileUploaderService={fileUploaderService}
+        isBlinking={attachPageBlinking}
+        isLoading={isLoading}
+        owner={owner}
+      />
+    </div>
   );
 };
